rename blue-button to health-records
