+++ src/js/health-records/reducers/modal.js
