+++ src/js/health-records/routes.jsx
@@ -1,10 +1,10 @@
-import BlueButtonApp from './containers/BlueButtonApp';
+import HealthRecordsApp from './containers/HealthRecordsApp';
 import Main from './containers/Main';
 import DownloadPage from './containers/DownloadPage';
 
 const routes = {
   path: '/',
-  component: BlueButtonApp,
+  component: HealthRecordsApp,
   indexRoute: { component: Main },
   childRoutes: [
     { path: 'download', component: DownloadPage },
