+++ src/sass/health-records/partials/_bb-modal.scss
