+++ test/health-records/containers/HealthRecordsApp.unit.spec.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import SkinDeep from 'skin-deep';
 import { expect } from 'chai';
 
-import { BlueButtonApp } from '../../../src/js/blue-button/containers/BlueButtonApp';
+import { HealthRecordsApp } from '../../../src/js/health-records/containers/HealthRecordsApp';
 
 const props = {
   modal: {
@@ -13,12 +13,11 @@ const props = {
   closeModal: () => {}
 };
 
-describe('<BlueButtonApp>', () => {
-  const tree = SkinDeep.shallowRender(<BlueButtonApp {...props}/>);
+describe('<HealthRecordsApp>', () => {
+  const tree = SkinDeep.shallowRender(<HealthRecordsApp {...props}/>);
 
   it('should render', () => {
     const vdom = tree.getRenderOutput();
     expect(vdom).to.exist;
   });
 });
-
